ai-services/requirements.txt
# Core AI/ML Libraries
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Optimization Libraries
ortools>=9.12.0
networkx>=3.1
cvxpy>=1.3.0

# Web Framework for AI APIs
flask>=2.3.0
flask-cors>=4.0.0
flask-restful>=0.3.10

# Configuration & Environment
python-dotenv>=1.0.0

# HTTP Requests
requests>=2.31.0
urllib3>=2.0.0

# Development & Testing
pytest>=7.4.0
pytest-flask>=1.2.0
black>=23.0.0
flake8>=6.0.0

# Data Visualization (optional)
matplotlib>=3.7.0
plotly>=5.15.0
seaborn>=0.12.0

# 3D Visualization and Bin Packing (optional)
py3dbp>=1.1.0
trimesh>=3.22.0

# Geographic and Routing (optional)
geopy>=2.3.0
shapely>=2.0.0
folium>=0.14.0

# Database Connectivity (optional)
psycopg2-binary>=2.9.0
pymongo>=4.4.0
redis>=4.6.0

# Data Processing Utilities
python-dateutil>=2.8.0
pytz>=2023.3

# Performance and Caching (optional)
joblib>=1.3.0
diskcache>=5.6.0

# Image Processing (optional)
opencv-python>=4.8.0
pillow>=10.0.0

# Testing Coverage
pytest-cov>=4.1.0

# Utilities
tqdm>=4.65.0
python-multipart==0.0.6
pydantic==2.0.3

# Development tools
black==23.7.0
flake8==6.0.0
mypy==1.4.1
